AnimationBehavior:
- tidy up based on code review comments
