+++ src/CommunityToolkit/Xamarin.CommunityToolkit.UnitTests/Behaviors/AnimationBehavior_Tests.cs
@@ -1,5 +1,4 @@
-﻿using System.Threading.Tasks;
-using NUnit.Framework;
+﻿using NUnit.Framework;
 using Xamarin.CommunityToolkit.Behaviors;
 using Xamarin.CommunityToolkit.UnitTests.Mocks;
 using Xamarin.Forms;
@@ -31,21 +30,7 @@ namespace Xamarin.CommunityToolkit.UnitTests.Behaviors
 
 			behavior.AnimateCommand.Execute(null);
 
-			Assert.That(mockAnimation.HasAnimated, Is.True);
-		}
-
-		class MockAnimationType : AnimationBase
-		{
-			public bool HasAnimated { get; private set; }
-
-			protected override uint DefaultDuration { get; set; } = 50;
-
-			public override Task Animate(View? view)
-			{
-				HasAnimated = true;
-
-				return Task.CompletedTask;
-			}
+			Assert.IsTrue(mockAnimation.HasAnimated);
 		}
 	}
 }
